# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "TPVV",
    "description" : "TPVV Ingeniería Web",
    "contact" : {
      "email" : "lac56@alu.ua.es"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/lac56-alu/TPVV/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://localhost:2000",
    "description" : "Desarrollo servidor basado en TPVV IW"
  }, {
    "url" : "",
    "description" : "Servidor"
  } ],
...
  "paths" : {
...
    "/tpvv/crearUsuario" : {
      "post" : {
        "tags" : [ "Usuario" ],
        "summary" : "Crea un nuevo usuario",
        "description" : "Podremos crear nuevos usuarios para nuestro sistema, será la única forma de introducir un nuevo cliente dentro del mismo.",
        "operationId" : "crearUsuario",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Usuario"
        },
        "responses" : {
          "201" : {
            "description" : "Usuario creado correctamente"
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/modificarUsuario" : {
      "put" : {
        "tags" : [ "Usuario" ],
        "summary" : "Podremos modificar un usuario",
        "description" : "Podremos modificar todos los campos de un usuario, es decir cambiar cada uno de los campos que ha introducido este para registrarse en nuestro sistema.",
        "operationId" : "modificarUsuario",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idUsuario",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/UsuarioModificar"
        },
        "responses" : {
          "200" : {
            "description" : "Detalles del usuario",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/borrarUsuario" : {
      "delete" : {
        "tags" : [ "Usuario" ],
        "summary" : "Podremos borrar un usuario en especifico",
        "description" : "Podremos borrar un usuario de nuestro sistema para siempre",
        "operationId" : "borrarUsuario",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idUsuario",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Usuario borrado con exito"
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/detallesUsuario" : {
      "get" : {
        "tags" : [ "Usuario" ],
        "summary" : "Obtener los detalles de un usuario en especifico",
        "description" : "Podremos consultar todos los detalles de un usuario, es decir visualizar cada uno de los campos que ha introducido este para registrarse en nuestro sistema.",
        "operationId" : "getDetallesUsuario",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idUsuario",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Detalles del usuario",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/listaUsuarios" : {
      "get" : {
        "tags" : [ "Usuario" ],
        "summary" : "Devuelve una lista con todos los usuarios",
        "description" : "Devuelve una lista con todos los usuarios que tenemos registrados dentro de nuestro sistema",
        "operationId" : "getListaUsuarios",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Lista con todos los usuarios",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/regenerarTokenUsuario" : {
      "post" : {
        "tags" : [ "Usuario" ],
        "summary" : "Regenerar el token del usuario",
        "description" : "Podremos volver a generar el token que va asociado al usuario, si por algún motivo este quiere realiar esta operación ya sea por seguridad o porque lo haya perdido.",
        "operationId" : "regenerarToken",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Nos devolverá el Token de ese usuario.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Usuario" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "nombre" : {
            "type" : "string",
            "example" : "Luis Alfonso"
          },
          "email" : {
            "type" : "string",
            "example" : "luis@gmail.com"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "example" : "aijsfbhausfbas"
          },
          "token" : {
            "type" : "string",
            "example" : "i989nasiudfbas54asd5as4da5s1d"
          },
          "nombreEmpresa" : {
            "type" : "string",
            "example" : "A reason"
          },
          "tipoUsuario" : {
            "type" : "string",
            "example" : "normal"
          }
        },
        "example" : {
          "password" : "aijsfbhausfbas",
          "tipoUsuario" : "normal",
          "id" : 1,
          "nombre" : "Luis Alfonso",
          "email" : "luis@gmail.com",
          "nombreEmpresa" : "A reason",
          "token" : "i989nasiudfbas54asd5as4da5s1d"
        }
      },
      "UsuarioModificar" : {
        "type" : "object",
        "properties" : {
          "nombre" : {
            "type" : "string",
            "example" : "Luis Alfonso"
          },
          "email" : {
            "type" : "string",
            "example" : "luis@gmail.com"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "example" : "aijsfbhausfbas"
          },
          "nombreEmpresa" : {
            "type" : "string",
            "example" : "A reason"
          }
        }
      },
...
      "inline_response_400" : {
        "type" : "object",
        "properties" : {
          "typeError" : {
            "type" : "string",
            "example" : "ERROR_PARAMETERS"
          },
          "messageError" : {
            "type" : "string",
            "example" : "Los parámetros introducidos son erroneos."
          }
        }
      },
      "inline_response_201" : {
        "type" : "object",
        "properties" : {
          "Token" : {
            "type" : "string",
            "example" : "jbHBSFBUIAKONaiaoizfcja0f09kasndjan"
          }
        },
        "example" : {
          "Token" : "jbHBSFBUIAKONaiaoizfcja0f09kasndjan"
        }
      },
...
      "inline_response_401" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_TOKEN"
          }
        }
      },
...
      "inline_response_404" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_RECURSO_NO_ENCONTRADO"
          }
        }
      },
...
      "inline_response_200_4" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Usuario"
            }
          }
        },
        "example" : {
          "data" : [ {
            "password" : "aijsfbhausfbas",
            "tipoUsuario" : "normal",
            "id" : 1,
            "nombre" : "Luis Alfonso",
            "email" : "luis@gmail.com",
            "nombreEmpresa" : "A reason",
            "token" : "i989nasiudfbas54asd5as4da5s1d"
          }, {
            "password" : "aijsfbhausfbas",
            "tipoUsuario" : "normal",
            "id" : 1,
            "nombre" : "Luis Alfonso",
            "email" : "luis@gmail.com",
            "nombreEmpresa" : "A reason",
            "token" : "i989nasiudfbas54asd5as4da5s1d"
          } ]
        }
      }
    },
...
    "requestBodies" : {
...
      "Usuario" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Usuario"
            }
          }
        },
        "required" : true
      },
      "UsuarioModificar" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/UsuarioModificar"
            }
          }
        },
        "required" : true
      },
...
    }
  }
}